Ported the entire system back to Windows using Visual Studio 2010 under Windows 7.  This entailed many changes, some of which will have broken the Linux and OSX builds, especially in subtle areas of POSIX compliance.
git-svn-id: https://s.snth.net/svn/rlk/thumb@2974 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ proj/panoview/sph-model.cpp
@@ -371,9 +371,10 @@ void sph_model::draw(const double *P, const double *V, const int *fv, int fc,
 
     glUseProgram(program);
     {
-        glUniform1f(glGetUniformLocation(program, "zoomk"), zoomk);
-        glUniform3f(glGetUniformLocation(program, "zoomv"),
-                    zoomv[0], zoomv[1], zoomv[2]);
+        glUniform1f(glGetUniformLocation(program, "zoomk"), GLfloat(zoomk));
+        glUniform3f(glGetUniformLocation(program, "zoomv"), GLfloat(zoomv[0]),
+														    GLfloat(zoomv[1]),
+															GLfloat(zoomv[2]));
 
         for (int i = 0; i < 6; ++i)
         {
@@ -475,7 +476,7 @@ void sph_model::set_fade(double k)
 {
     double t = k * k * (3.0 - 2.0 * k);
     glUseProgram(program);
-    glUniform1f(fader, t);
+    glUniform1f(fader, GLfloat(t));
 }
 
 static GLuint glGetUniformLocationv(GLuint program, const char *fmt, int d)
